+++ packages/okta-angular/CHANGELOG.md
@@ -1,3 +1,9 @@
+# Pending
+
+### Breaking Changes
+
+[#794](https://github.com/okta/okta-oidc-js/pull/794) `OktaAuthService.getUser` only uses `/userinfo` endpoint to retrieve user claims. `OktaAuthService.getUser` should be the recommended method to acquire user information as `idToken.claims` may become stale due to data updating.
+
 # 2.1.0
 
 ### Features